Create README.md
# Map_iOS

For identify-intersection branch, you can run the funciton in simulation. 

Once the simulator is up and running, you can click on the American Flag, and then hit the stop button to login the MapView.
Once you are in the map view use debug option in the menu bar to set the location manually. 
For custom location, I suggest you use the commented out location, which is the intersection of College Ave and Boston Ave.

Once you see the beacon in the MapView, you can then hit the button on the button of the simulator next to the 
plus sign. It manually tells the app to print out all attributes of that GeoLocation.